// frontend/src/entities/sentences/model/sentences.types.ts
export interface SentenceDataState {
	data: {
		nodes: Array<string>
		graph: Array<{
			first: Array<string> | string
			relation: string
			second: Array<string> | string
		}>
	}
	error?: {
		message: string
		word: string
	}
}

export enum SentenceDataActionsTokens {
	UPDATE_SENTENCE_TEXT_ACTION = 'UPDATE_SENTENCE_TEXT_ACTION',

	CLEAR_SENTENCE_PARSE_STATE_ACTION = 'CLEAR_SENTENCE_PARSE_STATE_ACTION',

	ASYNC_PARSE_SENTENCE_FROM_TEXT_ACTION = 'ASYNC_PARSE_SENTENCE_FROM_TEXT_ACTION',
	ASYNC_FIND_HY_TEXT_ACTION = 'ASYNC_FIND_HY_TEXT_ACTION',

	LOADING_PARSE_SENTENCE_FROM_TEXT_ACTION = 'LOADING_PARSE_SENTENCE_FROM_TEXT_ACTION',
	LOADED_PARSE_SENTENCE_FROM_TEXT_ACTION = 'LOADED_PARSE_SENTENCE_FROM_TEXT_ACTION',
	ERROR_PARSE_SENTENCE_FROM_TEXT_ACTION = 'ERROR_PARSE_SENTENCE_FROM_TEXT_ACTION'
}